requirements.txt
astroid==2.2.5
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
phonenumbers==8.10.19
Pillow==6.1.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
PySocks==1.7.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
twilio==6.31.0
typed-ast==1.4.0
urllib3==1.25.5
wrapt==1.11.2
